requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.0
PyJWT>=2.8.0
cryptography>=41.0.0
